kc-1 - webcam setup action renamed
